refactor(cli): move installFromGithub out of vessel.ts
installFromGithub serves ordinary `repo = "..."` GitHub deps and lived
in vessel.ts only for historical reasons. Move it (with its
downloadFromGithub helper) to cli/commands/install/install-from-github.ts
unchanged, ahead of dropping the rest of vessel support.
